// types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      chats: {
        Row: {
          chat_id: string
          created_at: string
          title: string | null
          updated_at: string
          user_id: string | null
        }
        Insert: {
          chat_id?: string
          created_at?: string
          title?: string | null
          updated_at?: string
          user_id?: string | null
        }
        Update: {
          chat_id?: string
          created_at?: string
          title?: string | null
          updated_at?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "chats_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["user_id"]
          },
        ]
      }
      logs: {
        Row: {
          created_at: string
          details: Json | null
          level: Database["public"]["Enums"]["log_level"]
          log_id: number
          message: string
          user_id: string
        }
        Insert: {
          created_at?: string
          details?: Json | null
          level?: Database["public"]["Enums"]["log_level"]
          log_id?: number
          message: string
          user_id?: string
        }
        Update: {
          created_at?: string
          details?: Json | null
          level?: Database["public"]["Enums"]["log_level"]
          log_id?: number
          message?: string
          user_id?: string
        }
        Relationships: []
      }
      messages: {
        Row: {
          chat_id: string
          content: string
          created_at: string
          generated_search_terms: Json | null
          message_id: string
          role: Database["public"]["Enums"]["message_role"]
          sources: Json | null
          web_search_required: boolean | null
        }
        Insert: {
          chat_id: string
          content: string
          created_at?: string
          generated_search_terms?: Json | null
          message_id?: string
          role: Database["public"]["Enums"]["message_role"]
          sources?: Json | null
          web_search_required?: boolean | null
        }
        Update: {
          chat_id?: string
          content?: string
          created_at?: string
          generated_search_terms?: Json | null
          message_id?: string
          role?: Database["public"]["Enums"]["message_role"]
          sources?: Json | null
          web_search_required?: boolean | null
        }
        Relationships: [
          {
            foreignKeyName: "messages_chat_id_fkey"
            columns: ["chat_id"]
            isOneToOne: false
            referencedRelation: "chats"
            referencedColumns: ["chat_id"]
          },
        ]
      }
      users: {
        Row: {
          avatar_url: string | null
          created_at: string
          email: string | null
          full_name: string | null
          updated_at: string
          user_id: string
        }
        Insert: {
          avatar_url?: string | null
          created_at?: string
          email?: string | null
          full_name?: string | null
          updated_at?: string
          user_id: string
        }
        Update: {
          avatar_url?: string | null
          created_at?: string
          email?: string | null
          full_name?: string | null
          updated_at?: string
          user_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      requesting_user_id: {
        Args: Record<PropertyKey, never>
        Returns: string
      }
    }
    Enums: {
      log_level: "DEBUG" | "INFO" | "WARN" | "ERROR" | "FATAL"
      message_role: "user" | "assistant"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never